requirements.txt
pyspark==3.2.1
pandas
kafka-python